updated contact views and added product model

+++ products/models.py
@@ -18,15 +18,13 @@ class Category(models.Model):
 
 
 class Product(models.Model):
-    category = models.ForeignKey(
-        "Category", null=True, blank=True, on_delete=models.SET_NULL
-    )
+    category = models.ForeignKey("Category", null=True, blank=True, on_delete=models.SET_NULL)
     sku = models.CharField(max_length=254, null=True, blank=True)
     name = models.CharField(max_length=254)
     description = models.TextField()
     price = models.DecimalField(max_digits=9, decimal_places=2)
     unique = models.BooleanField(default=True, blank="False")
-    quantity_available = models.PositiveBigIntegerField(default=1, blank=False)
+    quantity_available = models.PositiveIntegerField(default=1, blank=False)
     has_sizes = models.BooleanField(default=False, null=True, blank=True)
     dimensions = models.CharField(max_length=13, null=True, blank=True)
     rating = models.DecimalField(max_digits=6, decimal_places=2, null=True, blank=True)
